requirements.txt
Flask==3.0.2
praw==7.7.1
matplotlib==3.8.4
python-dotenv==1.0.1